hacked until docker-compose works
