+++ README.md
@@ -1,1 +1,2 @@
 Repository for the TOHacks 2019 Hackathon.
+Run `docker-compose up` to start a development environment
