+++ backend/package.json
@@ -11,7 +11,6 @@
   "dependencies": {
     "account": "^1.0.0",
     "bcryptjs": "^2.4.3",
-    "cors": "^2.8.5",
     "dotenv": "^8.0.0",
     "express": "^4.17.1",
     "levenshtein": "^1.0.5",
@@ -20,7 +19,7 @@
     "sequelize": "^5.8.12"
   },
   "devDependencies": {
-    "nodemon": "^1.19.1",
+    "cors": "^2.8.5",
     "sequelize-cli": "^5.5.0"
   }
 }
