+++ frontend/package.json
@@ -11,7 +11,9 @@
     "react-router-dom": "^5.0.1",
     "react-scripts": "0.9.5"
   },
-  "devDependencies": {},
+  "devDependencies": {
+    "npm-scripts": "^1.4.0"
+  },
   "scripts": {
     "start": "react-scripts start",
     "build": "react-scripts build",
